requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
dacite==0.0.13
dataclasses==0.6
emoji==2.1.0
frozenlist==1.3.1
google-api-core==2.10.1
google-api-python-client==2.64.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
google-cloud-texttospeech==2.12.1
googleapis-common-protos==1.56.4
googletrans==3.1.0a0
grpcio==1.49.0
grpcio-status==1.49.0
gspread==5.6.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httplib2==0.20.4
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.2
nest-asyncio==1.5.5
nextcord==2.2.0
nextcord-ext-menus==1.5.5
oauth2client==4.1.3
oauthlib==3.2.1
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymailtm==1.1.1
PyNaCl==1.4.0
pyparsing==3.0.9
pyperclip==1.8.2
pytz==2022.2.1
random-username==1.0.2
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
table2ascii==0.3.0
typing-extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.12
websockets==10.3
yarl==1.8.1
